% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_tracks}
\alias{load_tracks}
\title{Load Runkeeper(TM) .gpx files into a data.frame}
\usage{
load_tracks(gpxdir)
}
\arguments{
\item{gpxdir}{directory containing .gpx files and cardioActivities.csv extracted from a Runkeeper(TM) zip file.

Obtaining your data:

if{html}{figure{runkeeper_export.png}}}
}
\value{
data.frame of tracked data
}
\description{
Starting at \code{gpxdir}, loads all \code{.gpx} files and merges them
with CardioActivities.csv to produce a \code{data.frame} of all tracked
details.
}
\examples{
routes <- load_tracks(system.file("data", package="runkeepR"))
}

