% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_ggplot.R
\name{plot_ggplot}
\alias{plot_ggplot}
\title{Title}
\usage{
plot_ggplot(routes_all, filterType = NULL, trackPal = rainbow(7),
  center = NULL, zoom = 13)
}
\arguments{
\item{routes_all}{loaded and processed routes}

\item{filterType}{classification of routes to filter by (must be an existing type in the routes)}

\item{trackPal}{RColorBrewer palette name}

\item{center}{where to center the map}

\item{zoom}{zoom level (1:20; smaller number is tighter zoom)}
}
\value{
NULL (plots a ggplot map)
}
\description{
Title
}
\examples{
\dontrun{plot_ggplot(routes)}
}

