% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_leaflet.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{Plot the loaded routes using leaflet}
\usage{
plot_leaflet(routes_all, filterType = NULL, trackPal = rainbow(7), ...)
}
\arguments{
\item{routes_all}{loaded and processed routes}

\item{filterType}{classification of routes to filter by (must be an existing type in the routes)}

\item{trackPal}{RColorBrewer palette name}

\item{...}{other options to pass to addPolylines}
}
\value{
NULL (loads a leaflet map)
}
\description{
Plot the loaded routes using leaflet
}
\examples{
\dontrun{plot_leaflet(routes, filterType="Walking", trackPal="Accent")}
}

