% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_leaflet.R
\name{points_to_line}
\alias{points_to_line}
\title{Convert points to lines}
\usage{
points_to_line(data, long, lat, id_field = NULL, sort_field = NULL)
}
\arguments{
\item{data}{the data}

\item{long}{longitude}

\item{lat}{latitude}

\item{id_field}{id field}

\item{sort_field}{sort field}
}
\value{
lines object (?)
}
\description{
Convert points to lines
}
\references{
\url{https://rpubs.com/walkerke/points_to_line}
}
\keyword{Internal}

