% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_RK_GPX}
\alias{read_RK_GPX}
\title{Read a Runkeeper(TM) \code{.gpx} file into a data.frame}
\usage{
read_RK_GPX(gpxfile)
}
\arguments{
\item{gpxfile}{name of the .gpx file to process}
}
\value{
data.frame of track details containing all available details for the tracked routes
}
\description{
Raw read of a single \code{.gpx} file. Only applicable to
Runkeeper(TM) \code{.gpx} files. Exported in case it is of use.
}
\details{
For a more general use version see \code{plotKML::readGPx} which fails to
read in elevation from Runkeeper(TM) files.
}
\examples{
routes <- load_tracks(system.file("data", package="runkeepR"))
}

