\name{runkeepR-package}
\alias{runkeepR-package}
\alias{runkeepR}
\docType{package}
\title{
\packageTitle{runkeepR}
}
\description{
\packageDescription{runkeepR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{runkeepR}
\packageIndices{runkeepR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{runkeepR}

Maintainer: \packageMaintainer{runkeepR}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }

