% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{summarise_runs}
\alias{summarise_runs}
\title{Show a summary of the Runkeeper(TM) data}
\usage{
summarise_runs(rundata, by = "trkname", dashboard = TRUE)
}
\arguments{
\item{rundata}{the data to process}

\item{by}{grouping category. Default is \code{trkname}; the unique tracks.}

\item{dashboard}{present the information in a shinydashboard?}
}
\description{
Show a summary of the Runkeeper(TM) data
}

