% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_ggplot.R
\name{plot_ggplot}
\alias{plot_ggplot}
\title{Plot Runkeeper(TM) route data using ggplot2}
\usage{
plot_ggplot(routes_all, filterType = NULL, trackPal = rainbow(7),
  center = NULL, zoom = 13)
}
\arguments{
\item{routes_all}{loaded and processed routes}

\item{filterType}{classification of routes to filter by (must be an existing type in the routes)}

\item{trackPal}{RColorBrewer palette name}

\item{center}{where to center the map. Defaults to the median lat/lon from the filtered data.}

\item{zoom}{zoom level (1:20; smaller number is tighter zoom)}
}
\value{
NULL (plots a ggplot map)
}
\description{
This function performs several tasks to produce a layered map of routes. A map, centered
on the median lat/lon (most common location in the filtered data) at the requested zoom level
is obtained from Google Maps via the maps API. Routes are then added one by one in
distinct colours.
}
\examples{
\dontrun{plot_ggplot(routes, filterType="Walking", trackPal=rainbow(8), zoom=14)}
}

