% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_RK_GPX}
\alias{read_RK_GPX}
\title{Read a Runkeeper(TM) \code{.gpx} file into a data.frame}
\usage{
read_RK_GPX(gpxfile)
}
\arguments{
\item{gpxfile}{name of the .gpx file to process}
}
\value{
data.frame of track details containing all available details for the tracked routes  
with additional class \code{runkeepR_data}
}
\description{
Raw read of a single \code{.gpx} file. Only applicable to 
Runkeeper(TM) \code{.gpx} files. Exported in case it is of use.
}
\details{
For a more general use version see \code{plotKML::readGPx} which fails to 
read in elevation from Runkeeper(TM) files.
}
\examples{
## load test data distributed with this package
## test data is a single track; 
## a hike around Anstey Hill, Adelaide, South Australia
## https://en.wikipedia.org/wiki/Anstey_Hill_Recreation_Park

library(runkeepR)
routes <- load_tracks(system.file("extdata", package="runkeepR"))
class(routes) 
## [1] "runkeepR_data" "data.frame"  

}

