\name{runkeepR-package}
\alias{runkeepR-package}
\alias{runkeepR}
\docType{package}
\title{
\packageTitle{runkeepR}
}
\description{
\packageDescription{runkeepR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{runkeepR}
\packageIndices{runkeepR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{J. Carroll}

Maintainer: \packageMaintainer{J. Carroll}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
